refactor: separate authorize logics
