+++ app/ping/api.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends
 
-from app.core.auth.dependency import authenticated_user
+from app.core.auth.access import authenticated_user
 from app.db.tables import User
 
 app_router = APIRouter()
